requirements.txt
# Core dependencies
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
click==8.1.7
rich==13.5.2
python-dotenv>=1.0.0
gitpython==3.1.32
pydantic==2.5.3

# API and Web
fastapi==0.109.0
uvicorn==0.25.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
black==23.12.1
isort==5.13.2
pylint==3.0.3

# Jinja2
jinja2>=3.1.2

# Requests
requests>=2.31.0

# YAML
pyyaml>=6.0.1 